refactor(cli): use correct paths to import modules
